+++ src/components/base/ImageUpload/index.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react';
-import axios from 'axios';
+import { uploadProductImage } from '../../../actions/product';
 
 // https://codepen.io/hartzis/pen/VvNGZP
 class ImageUpload extends Component {
@@ -12,44 +12,9 @@ class ImageUpload extends Component {
     e.preventDefault();
     // TODO: do something with -> this.state.file
     //console.log('handle uploading-', this.state.file);
-    let AWS_SIGNURL = 'https://78n3id557f.execute-api.us-east-1.amazonaws.com/api/product/image/signurl';
     let { productId } = this.props;
-    console.log('productId:', productId);
-
     let selectedFile = this.state.file;
-    let { name: fileName, type: fileType } = selectedFile;
-    axios.post(AWS_SIGNURL,{
-      fileName : fileName,
-      fileType : fileType
-    })
-    .then(response => {
-        let returnData = response.data.returnData;
-        let signedRequest = returnData.signedRequest;
-        let url = returnData.url;
-        this.setState({
-            fileUpload: { success: true, url: url }
-        });
-        
-        //console.log("Recieved a signed request " + signedRequest);
-        let options = {
-            headers: {
-              'Content-Type': fileType
-            }
-        };
-        axios.put(signedRequest, selectedFile, options)
-        .then(result => {
-            console.log("Response from s3:", result);
-            this.setState({
-                fileUpload: { success: true, url: url }
-            });
-        })
-        .catch(error => {
-            alert("ERROR " + JSON.stringify(error));
-        });
-    })
-    .catch(error => {
-        alert("ERROR " + JSON.stringify(error));
-    });
+    return uploadProductImage(selectedFile, productId);
   }
   
   _handleImageChange(e) {
